requirements.txt
asgiref==3.7.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.0.1
crispy-bootstrap5==2023.10
cryptography==39.0.1
defusedxml==0.7.1
dj-database-url==2.2.0
dj-email-url==1.0.6
django==5.0.3
django-allauth==0.61.1
django-cache-url==3.4.5
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
environs==11.0.0
gunicorn==21.2.0
idna==3.4
marshmallow==3.22.0
oauthlib==3.2.2
packaging==23.1
pillow==10.4.0
psycopg==3.1.18
psycopg-binary==3.1.18
pycparser==2.21
pyjwt==2.6.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
typing-extensions==4.9.0
urllib3==1.26.14
whitenoise==6.6.0
